Add `copy-object` method
[closes #1]

Added new method as specified in the related issue. For now there are no used
extra options for this method, so that is why we just check we get a map as
input and nothing else. We provide this optional argument for the future, as
we will probably need it.

We only support copying objects between same storage system for now, in order
to keep things simple as there is no such a use-case for it yet.

Aligned docstrings of the other methods there.

Prepared code for new release, as it was not done before.